Moved redux connect to separate files
