+++ src/components/index.js
@@ -1,9 +1,9 @@
-export {default as App} from "./app/app";
+export {default as App} from "./app/app.connect";
 export {default as Main} from "./main/main";
-export {default as MoviesList} from "./movies-list/movies-list";
-export {default as MovieCard} from "./movie-card/movie-card";
+export {default as MoviesList} from "./movies-list/movies-list.connect";
+export {default as MovieCard} from "./movie-card/movie-card.connect";
 export {default as MovieDetails} from "./movie-details/movie-details";
 export {default as VideoPlayer} from "./video-player/video-player";
 export {default as Tabs} from "./tabs/tabs";
 export {default as Tab} from "./tab/tab";
-export {default as GenresList} from "./genres-list/genres-list";
+export {default as GenresList} from "./genres-list/genres-list.connect";
