+++ src/components/movie-card/movie-card.e2e.test.js
@@ -1,4 +1,4 @@
-import {MovieCard} from "./movie-card";
+import MovieCard from "./movie-card";
 import {generateId} from "@utils";
 
 const HANDLE_HOVER_CALL_COUNT = 1;
