+++ src/components/movies-list/movies-list.test.js
@@ -1,4 +1,4 @@
-import {MoviesList} from "./movies-list";
+import {MoviesList} from "@components";
 import {generateId} from "@utils";
 import {createStore} from "redux";
 import {Provider} from "react-redux";
